Fix the comparison in test_compatibility for NumPy 2.0.0
The comparison in `mpmath/tests/test_convert.py::test_compatibility`
failed for `np.float16` in NumPy 2.0.0 since `2.0**-53` cannot be
represented in half-precision floating point type.  (NumPy 2
implicitly converts the second operand to the type of the first one.)

Explicitly convert it to `np.float64` to ensure that the comparison is
done in sufficiently precise type.

// edited by Sergey B Kirpichev

+++ mpmath/tests/test_convert.py
@@ -259,7 +259,7 @@ def test_compatibility():
         # Handle the weird types
         try: diff = np.abs(type(np.sqrt(x))(sqrt(x)) - np.sqrt(x))
         except: continue
-        assert diff < 2.0**-53
+        assert diff < np.float64(2.0**-53)
     assert mpf(np.float64('inf')) == inf
     assert isnan(mp.npconvert(np.float64('nan')))
     if hasattr(np, "float128"):
